requirements.txt
Babel==2.3.4
PySocks==1.6.5
PyYAML==3.12
SQLAlchemy==1.1.5
Unidecode==0.04.20
agate==1.5.5
agate-dbf==0.2.0
agate-excel==0.2.0
agate-sql==0.5.0
argparse==1.2.1
awesome-slugify==1.6.5
dbfread==2.0.7
et-xmlfile==1.0.1
future==0.16.0
googleads==5.0.0
httplib2==0.9.2
isodate==0.5.4
jdcal==1.3
leather==0.3.3
oauth2client==1.5.2
openpyxl==2.4.2
parsedatetime==2.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pytimeparse==1.1.6
pytz==2016.10
regex==2017.02.08
rsa==3.4.2
six==1.10.0
suds-jurko==0.6
wsgiref==0.1.2
xlrd==1.0.0
xmltodict==0.10.2
